Extract utility functions to make them more easily unit testable (without requiring all the dependencies in cspm.py); Add unit tests for new methods
